Simple search bar, currently only takes in props and uses those to reference user input.
TODO: Plug the prop passed in either into an api call to our backend, or
referencing a global state array that holds the name of all the games.
I'm thinking I'll create a worker that will update the state with all
games in the events db every 60 seconds or so to minimize calls to our
back end and reduce moving parts.