Big push, but it's mostly a refactoring
I ripped up all the old components and spread them over the current
functionality. The end result is hopefully more, smaller components that
are easier to understand on their own. The functionality is all
returned, and hopefully more stable.

Notes: This may not be the ultimate refactoring. I had to make some
concessions to keep certain states at the lowest level. In other cases
there are props handed down through 3-4 components to keep the lower
ends as simple as possible.